Port BertEmbeddings from Hugging Face transformers
Also:

- Update the LayerNorm layer to use a VarStore.
- Add an Embedding layer. We cannot access everything we need in
  tch:nn:Embedding.

The forward pass of BertEmbeddings is tested by applying it to some
input and validating the output against output of the Hugging Face
transformers package. The test is behind the 'model-tests' feature
gate and requires 'bert-base-german-cased.hdf5' to be in the testdata
directory. A future PR will add a script to automatically create this
file.