Add support for attention masks in BertEncoder
Normally, we want to mask attention for inactive timesteps, so that
their hidden states do not leak into the representations of active
time steps.

The API differs from Hugging Face transformers. In Hugging Face
transformers, the mask is converted to appropriate values that result
in near-zero probabilities in BertModel. This means that BertEncoder
already expects this special representation. However, since
BertEncoder is the typical 'entry point' in sticker-transformers, we
perform the conversion in BertEncoder.

The forward methods of the layer types that an encoder uses accept the
LogitsMask type, which can only be constructed from a bool tensor.